Update README.md
This software scans your machine for popular softwares and generates QR codes for the bitcoin addresses provided by the original developers so that you can donate bitcoin to the software developers.

Please donate bitcoin to developers. There is so much that they give us, and so freely. They have bills, too, you know.